+++ dart/dynamics/BodyNode.h
@@ -220,13 +220,13 @@ public:
   BodyNode* getChildBodyNode(int _index) const;
 
   /// Return the number of child bodynodes
-  int getNumChildBodyNodes() const;
+  size_t getNumChildBodyNodes() const;
 
   /// Add a marker into the bodynode
   void addMarker(Marker* _marker);
 
   /// Return the number of markers of the bodynode
-  int getNumMarkers() const;
+  size_t getNumMarkers() const;
 
   /// Return _index-th marker of the bodynode
   Marker* getMarker(int _index) const;
@@ -235,11 +235,11 @@ public:
   bool dependsOn(int _genCoordIndex) const;
 
   /// The number of the generalized coordinates by which this node is affected
-  int getNumDependentGenCoords() const;
+  size_t getNumDependentGenCoords() const;
 
   /// Return a generalized coordinate index from the array index
   /// (< getNumDependentDofs)
-  int getDependentGenCoordIndex(int _arrayIndex) const;
+  size_t getDependentGenCoordIndex(int _arrayIndex) const;
 
   //--------------------------------------------------------------------------
   // Properties updated by dynamics (kinematics)
